+++ src/Lykke.Service.OperationsRepository.AzureRepositories/CashOperations/TransferEventsRepository.cs
@@ -40,7 +40,7 @@ namespace Lykke.Service.OperationsRepository.AzureRepositories.CashOperations
             set { StateField = value.ToString(); }
         }
         public string ClientId { get; set; }
-        public decimal FeeSize { get; set; }
+        public double FeeSize { get; set; }
         public FeeType FeeType { get; set; }
         public string FeeTypeText
         {
